Add ORBIT list functionality to Findit
This is a hidden feature that won't affect existing functionality in
Findit. Go ahead and deploy it if you need to.

ORBIT lists are used by the research registration team to catch
publications that ORBIT's automatic searches don't catch. The lists are
currently being created by manual runs done by yours truly (suck suck).

This commit adds an interface for defining and managing query strings
and for viewing and managing the resulting documents that might need
to be registered in ORBIT.

There is also a new rake task for doing the actual searches which
should probably be run once a day by a cron job or similar.